Added HOc for not auth paths
