+++ client/src/hoc/WithNotAuth.js
@@ -1,0 +1,5 @@
+import useNotAuth from '../hooks/useNotAuth';
+
+const WithNotAuth = (props) => !useNotAuth(props) && props.children;
+
+export default WithNotAuth;
